+++ packages/melt/src/lib/builders/index.ts
@@ -10,6 +10,8 @@ export * from "./Toggle.svelte";
 export * from "./Slider.svelte";
 export * from "./Tree.svelte";
 export * from "./utils.svelte";
+export * from "./FileUpload.svelte";
+
 export * from "./Tooltip.svelte";
 export * from "./Tree.svelte";
 export * from "./Avatar.svelte";