+++ packages/melt/src/lib/components/Avatar.svelte
@@ -1,17 +1,17 @@
 <script lang="ts">
-	import { Avatar, type AvatarProps, type ImageLoadingStatus } from "../builders/Avatar.svelte";
+	import { Avatar as Builder, type AvatarProps, type ImageLoadingStatus } from "../builders/Avatar.svelte";
 	import { type Snippet } from "svelte";
 	import type { ComponentProps } from "../types";
 	import { getters } from "$lib/builders";
 
 	type Props = ComponentProps<AvatarProps> & {
-		children: Snippet<[Avatar]>;
+		children: Snippet<[Builder]>;
 		onLoadingStatusChange?: (value: ImageLoadingStatus) => void | undefined;
 	};
 
 	let { children, onLoadingStatusChange, ...rest }: Props = $props();
 
-	const avatar = new Avatar(getters({ ...rest }));
+	export const avatar = new Builder(getters({ ...rest }));
 </script>
 
 {@render children(avatar)}