+++ packages/melt/src/lib/components/Slider.svelte
@@ -1,16 +1,16 @@
 <script lang="ts">
-	import { Slider, type SliderProps } from "../builders/Slider.svelte";
+	import { Slider as Builder, type SliderProps } from "../builders/Slider.svelte";
 	import { type Snippet } from "svelte";
 	import type { ComponentProps } from "../types";
 	import { getters } from "../builders/utils.svelte";
 
 	type Props = ComponentProps<SliderProps> & {
-		children: Snippet<[Slider]>;
+		children: Snippet<[Builder]>;
 	};
 
 	let { value = $bindable(), children, ...rest }: Props = $props();
 
-	const slider = new Slider({
+	export const slider = new Builder({
 		value: () => value,
 		onValueChange: (v) => (value = v),
 		...getters(rest),