+++ packages/melt/src/lib/components/Tabs.svelte
@@ -1,16 +1,16 @@
 <script lang="ts">
-	import { Tabs, type TabsProps } from "../builders/Tabs.svelte";
+	import { Tabs as Builder, type TabsProps } from "../builders/Tabs.svelte";
 	import { type Snippet } from "svelte";
 	import type { ComponentProps } from "../types";
 	import { getters } from "../builders/utils.svelte";
 
 	type Props = ComponentProps<TabsProps> & {
-		children: Snippet<[Tabs]>;
+		children: Snippet<[Builder]>;
 	};
 
 	let { value = $bindable(), children, ...rest }: Props = $props();
 
-	const tabs = new Tabs({
+	export const tabs = new Builder({
 		value: () => value,
 		onValueChange: (v) => (value = v),
 		...getters(rest),